requirements.txt
altgraph==0.17
fbs==0.9.0
future==0.18.2
macholib==1.14
numpy==1.19.2
pefile==2019.4.18
Pillow==7.2.0
PyInstaller==3.4
PyQt5==5.15.1
PyQt5-sip==12.8.1
pywin32-ctypes==0.2.0
